Personal_project/tiktok_scheduler/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
SQLAlchemy==2.0.20
cryptography==41.0.4
python-dotenv==1.0.0
requests==2.31.0
APScheduler==3.10.4
gunicorn==21.2.0
bcrypt==4.0.1
pillow==10.4.0
pycryptodome==3.19.0
jsonschema==4.20.0